Consolidate Service Extension methods
